requirements.txt
numpy==1.19.5
pandas==1.1.5
matplotlib==3.3.4
openpyxl==3.0.9
xlsxwriter==3.0.2
Pillow==8.4.0
cx_Freeze==6.8.4
lief==0.11.5
importlib_metadata==4.13.0